pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "grpcio-tools >=1.42,<1.45",
    "PyYaml >=5.3.1,<6.1",
    "types-psutil ~=5.9",
]
build-backend = "setuptools.build_meta"

[project]
name = "cogment"
description = "Cogment python SDK"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dynamic = ["readme", "version"]
requires-python = ">=3.7, <3.11"
dependencies = [
    "grpcio >=1.42, <1.45",
    "grpcio-reflection >=1.42, <1.45",
    "protobuf ~=3.7",
    "prometheus-client >= 0.8, <0.9",
    "typing_extensions >= 4.6.3, < 4.7.0",
]

[[project.authors]]
name = "Artificial Intelligence Redefined"
email = "dev+cogment@ai-r.com"

[project.urls]
Homepage = "https://cogment.ai"
repository = "https://github.com/cogment/cogment-py-sdk"
"Reference Documentation" = "https://cogment.ai/docs/reference/python"

[project.optional-dependencies]
generate = [
    "PyYaml >=5.3.1,<6.1",
    "psutil ~=5.9",
    "grpcio-tools ==1.43",
    "click ~= 8.0.3",
]

[tool.setuptools]
packages = ["cogment", "cogment.api"]
include-package-data = false

[tool.setuptools.package-data]
"*" = ["*.proto"]

[tool.setuptools.data-files]
"." = ["CHANGELOG.md"]

[tool.setuptools.dynamic.readme]
file = ["README.md", "CHANGELOG.md"]
content-type = "text/markdown"

[tool.setuptools.dynamic.version]
attr = "cogment.version.__version__"

[tool.pytest.ini_options]
log_level = "DEBUG"
log_cli = "True"
# The log format is setup to mirror cogment's
log_format = "%(asctime)s [%(levelname)s] %(message)s"
log_date_format = "%Y-%m-%dT%H:%M:%S%z"
addopts = "-rfs --capture=no --junit-xml=integration_test_report.xml"
junit_suite_name = "cogment_py_sdk_tests"
timeout = 15

[tool.mypy]
python_version = 3.7
warn_return_any = true
warn_unused_configs = true
strict_optional = false
exclude = ['tests', 'setup.py']

[[tool.mypy.overrides]]
module = ["grpc", "prometheus_client"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["cogment.api.*"]
ignore_errors = true

[tool.black]
line-length = 120
target-version = ["py39"]
color = true
